chatbot_py/requirements.txt
# chatbot_py/requirements.txt

fastapi
uvicorn[standard]
google-generativeai
python-dotenv
requests 
pydantic